requirements.txt

flask==2.2.3
werkzeug==2.2.3
pandas>=1.5.3
numpy>=1.23.5
requests==2.26.0
google-cloud-storage==2.1.0
gunicorn==20.1.0